Move SyncActionError to separate dir
