+++ src/scripts/modules/wr-google-drive/react/components/FileModal.jsx
@@ -6,7 +6,7 @@ import InputTab from './InputTab';
 import FileTab from './FileTab';
 import ActionTab from './ActionTab';
 import StorageTablesStore from '../../../components/stores/StorageTablesStore';
-import SyncActionError from '../../../../utils/SyncActionError';
+import SyncActionError from '../../../../utils/errors/SyncActionError';
 
 export default React.createClass({
   propTypes: {
